LittleLemon/requirements.txt
django 
djangorestframework
